+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Template Overview - blocked out page for component builder assistance [#75]
 - Template Edit Section  - blocked out page for component builder assistance [#75]
 - Template Edit Question - blocked out page for component builder assistance [#75]
+- Set up a directory for the account pages, and added the Connections page [#128]
+- Added new shared LeftSidebar, TooltipWithDialog, and ButtonWithImage components. Updated styleguide with the new TooltipWithDialog and ButtonWithImage components. [#128]
 - Fixed type bugs in the template history page which was breaking the build.
 - Added csrf token to header for /login and /signup page and error handling to refresh csrf and auth tokens [#116]
 - Added cypress for functional tests. Added a functional test to check for HTTP-only auth cookies [#116]